clk: meson: Export a1 functions to common file
Export several useful functions from a1 to a common file. These will be
used for s4 support.

Patch-cc: Lukasz Majewski <lukma@denx.de>
